+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added variants to Product element’s GraphQL schema definition.
 - It’s now possible to query for products and variants by their custom field values via GraphQL.
 - It’s now possible to query for orders by multiple email addresses. ([#1361](https://github.com/craftcms/commerce/issues/1361))
+- Added `craft\commerce\service\Variants::getVariantGqlContentArguments()`.
 
 ### Changed
 - `craft\commerce\controllers\CartController::$_cart` and `craft\commerce\controllers\CartController::$_cartVariable` to protected properties.